게시글 삭제, 수정에서 글이 없는 경우 예외처리 구현

+++ src/com/sbs/example/board/App.java
@@ -77,6 +77,19 @@ public class App {
 		} else if (cmd.startsWith("article modify")) {
 
 			int id = Integer.parseInt(cmd.split(" ")[2].trim());
+			
+			SecSql sql = new SecSql();
+			
+			sql.append("SELECT COUNT(*)");
+			sql.append("FROM article");
+			sql.append("WHERE id = ?", id);
+			
+			int articlesCount = DBUtil.selectRowIntValue(conn, sql);
+			
+			if(articlesCount == 0) {
+				System.out.printf("%d번 게시글이 존재하지 않습니다.\n", id);
+				return 0;
+			}
 
 			String title;
 			String body;
@@ -87,7 +100,7 @@ public class App {
 			System.out.printf("새 내용: ");
 			body = scanner.nextLine();
 
-			SecSql sql = new SecSql();
+			sql = new SecSql();
 			
 			sql.append("UPDATE article");
 			sql.append("SET regDate = NOW()");
@@ -133,8 +146,21 @@ public class App {
 			int id = Integer.parseInt(cmd.split(" ")[2].trim());
 			
 			System.out.println("== 게시글 삭제 ==");
-
+			
 			SecSql sql = new SecSql();
+			
+			sql.append("SELECT COUNT(*)");
+			sql.append("FROM article");
+			sql.append("WHERE id = ?", id);
+			
+			int articlesCount = DBUtil.selectRowIntValue(conn, sql);
+			
+			if(articlesCount == 0) {
+				System.out.printf("%d번 게시글이 존재하지 않습니다.\n", id);
+				return 0;
+			}
+
+			sql = new SecSql();
 			
 			sql.append("DELETE FROM article");
 			sql.append("WHERE id = ?", id);
